requirements.txt
aiohttp==3.10.5
beautifulsoup4==4.12.3
python-telegram-bot==20.8
aiosqlite==0.20.0